LayoutTests/http/tests/security/contentSecurityPolicy/media-src-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/resources/load-and-stall.cgi?name=../../../media/content/test.mp4&mimeType=video/mp4&stallAt=100000 because it does not appear in the media-src directive of the Content Security Policy.

END OF TEST
This test passes if it doesn't alert failure.
